+++ packages/contracts/contracts/libs/WalletUtils.sol
@@ -55,8 +55,14 @@ library WalletUtils {
                 allPis[index][4] = op.joinSplitTxs[j].nullifierA;
                 allPis[index][5] = op.joinSplitTxs[j].nullifierB;
                 allPis[index][6] = digests[i];
-                allPis[index][7] = op.joinSplitTxs[j].encodedAssetAddr;
-                allPis[index][8] = op.joinSplitTxs[j].encodedAssetId;
+                allPis[index][7] = op
+                    .joinSplitTxs[j]
+                    .encodedAsset
+                    .encodedAssetAddr;
+                allPis[index][8] = op
+                    .joinSplitTxs[j]
+                    .encodedAsset
+                    .encodedAssetId;
                 index++;
             }
         }
@@ -96,8 +102,8 @@ library WalletUtils {
                         op.joinSplitTxs[i].newNoteACommitment,
                         op.joinSplitTxs[i].newNoteBCommitment,
                         op.joinSplitTxs[i].publicSpend,
-                        op.joinSplitTxs[i].encodedAssetAddr,
-                        op.joinSplitTxs[i].encodedAssetId
+                        op.joinSplitTxs[i].encodedAsset.encodedAssetAddr,
+                        op.joinSplitTxs[i].encodedAsset.encodedAssetId
                     )
                 )
             );
@@ -165,6 +171,7 @@ library WalletUtils {
 
     // TODO: properly process this failure case
     function _unsuccessfulOperation(
+        Operation calldata op,
         bytes memory result
     ) internal pure returns (OperationResult memory) {
         bool[] memory callSuccesses = new bool[](1);
@@ -179,18 +186,16 @@ library WalletUtils {
                 callResults: callResults,
                 executionGas: 0,
                 verificationGas: 0,
-                numRefunds: 0
+                numRefunds: op.joinSplitTxs.length +
+                    op.encodedRefundAssets.length
             });
     }
 
     function _verificationGasForOp(
-        Bundle calldata bundle,
-        uint256 opIndex,
-        uint256 batchVerificationGas
+        Operation calldata op,
+        uint256 perJoinSplitGas
     ) internal pure returns (uint256) {
-        return
-            (batchVerificationGas / bundle.totalNumJoinSplits()) *
-            bundle.operations[opIndex].joinSplitTxs.length;
+        return perJoinSplitGas * op.joinSplitTxs.length;
     }
 
     function _calculateBundlerGasAssetPayout(